+++ keg_apps/db/model/entities.py
@@ -1,3 +1,5 @@
+from blazeutils.strings import randchars
+
 from keg.db import db
 
 
@@ -6,6 +8,13 @@ class Blog(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.Unicode(50), unique=True, nullable=False)
+
+    @classmethod
+    def testing_create(cls):
+        blog = Blog(title=randchars())
+        db.session.add(blog)
+        db.session.commit()
+        return blog
 
 
 ####
@@ -28,6 +37,13 @@ class PGDud2(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.Unicode(50), unique=True, nullable=False)
+
+    @classmethod
+    def testing_create(cls):
+        dud = cls(name=randchars())
+        db.session.add(dud)
+        db.session.commit()
+        return dud
 
 
 class SADud(db.Model):